// socialmedia-backend/src/utils/types.ts
export enum ProfileType {
  ADMIN = "ADMIN",
  USER = "USER",
}

export enum Status_AD {
  PUBLISH = "PUBLISH",
  SCHEDULE = "SCHEDULE",
  TEMPORARY = "TEMPORARY",
}

export enum SocialPlatform {
  FACEBOOK = "FACEBOOK",
  INSTAGRAM = "INSTAGRAM",
  TWITTER = "TWITTER",
}

export enum AdvertisementType {
  PAGE = "PAGE",
  ACCOUNT = "ACCOUNT",
  TWITTER = "TWITTER",
}